requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-filter==23.4
celery==5.3.4
python-dotenv==1.0.0
Pillow==10.1.0
